Move entities, errors folder
