+++ website/data-storage/src/common/errors/interfaces.ts
