+++ website/data-storage/src/consumer/types.ts
@@ -1,3 +1,3 @@
-import { ExecutionError } from 'src/errors/interfaces';
+import { ExecutionError } from 'src/common/errors/interfaces';
 
 export type Result<T> = Promise<T | ExecutionError>;
