+++ website/data-storage/src/program/program.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { Program } from '../entities/program.entity';
+import { Program } from '../database/entities/program.entity';
 import { ProgramService } from './program.service';
 import { ProgramRepo } from './program.repo';
