+++ website/data-storage/src/program/program.repo.ts
@@ -2,7 +2,7 @@ import { Repository } from 'typeorm';
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 
-import { Program } from '../entities';
+import { Program } from '../database/entities';
 import { GetAllProgramsParams, GetAllUserProgramsParams } from '@gear-js/common';
 import { sqlWhereWithILike } from '../utils/sql-where-with-ilike';
 import { PAGINATION_LIMIT } from '../config/configuration';
